Convert ServiceExtensions to static class and add methods
The `ServiceExtensions` class in `ServiceExtensions.cs` has been
converted from a regular class to a static class to allow it to
contain extension methods.

Two new extension methods have been added to the `ServiceExtensions`
class:
- `ConfigureCors`: Configures CORS by adding a policy named
  "CorsPolicy" that allows any origin, method, and header.
- `ConfigureIISIntegration`: Configures IIS integration with default
  options.

In `Program.cs`, the `CompanyEmployees.Extensions` namespace has been
imported to use the newly added extension methods. The `ConfigureCors`
and `ConfigureIISIntegration` methods are called on the `builder.Services`
object to add the respective configurations to the service collection.